crawlWeb/ improve the recording of information about the site and its pages in the database tables. / Implement updating 'status-time' and 'status' in site table while crawling /

+++ src/main/java/searchengine/repositories/SiteRepository.java
@@ -3,5 +3,8 @@ package searchengine.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import searchengine.model.SiteModel;
 
-public interface SiteRepository extends JpaRepository<SiteModel, Long> {
+public interface SiteRepository extends JpaRepository<SiteModel, Integer> {
+
+
+
 }
